+++ src/pageEditor/tabs/editTab/dataPanel/FoundationDataPanel.tsx
@@ -21,8 +21,6 @@ import { UUID } from "@/core";
 import { useSelector } from "react-redux";
 import { makeSelectBlockTrace } from "@/pageEditor/slices/runtimeSelectors";
 import { Nav, Tab } from "react-bootstrap";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faInfoCircle } from "@fortawesome/free-solid-svg-icons";
 import dataPanelStyles from "@/pageEditor/tabs/dataPanelTabs.module.scss";
 import ExtensionPointPreview from "@/pageEditor/tabs/effect/ExtensionPointPreview";
 import useDataPanelActiveTabKey from "@/pageEditor/tabs/editTab/dataPanel/useDataPanelActiveTabKey";
@@ -31,6 +29,8 @@ import { FormState } from "@/pageEditor/pageEditorTypes";
 import PageStateTab from "./PageStateTab";
 import { DataPanelTabKey } from "@/pageEditor/tabs/editTab/dataPanel/dataPanelTypes";
 import DataTabJsonTree from "./DataTabJsonTree";
+import StateTab from "./tabs/StateTab";
+import ConfigurationTab from "./tabs/ConfigurationTab";
 
 const FoundationDataPanel: React.FC<{
   firstBlockInstanceId?: UUID;
@@ -59,7 +59,7 @@ const FoundationDataPanel: React.FC<{
         {showDeveloperTabs && (
           <>
             <Nav.Item className={dataPanelStyles.tabNav}>
-              <Nav.Link eventKey={DataPanelTabKey.Formik}>Formik</Nav.Link>
+              <Nav.Link eventKey={DataPanelTabKey.Formik}>State</Nav.Link>
             </Nav.Item>
             <Nav.Item className={dataPanelStyles.tabNav}>
               <Nav.Link eventKey={DataPanelTabKey.BlockConfig}>
@@ -93,39 +93,8 @@ const FoundationDataPanel: React.FC<{
         </Tab.Pane>
         {showDeveloperTabs && (
           <>
-            <Tab.Pane
-              eventKey={DataPanelTabKey.Formik}
-              className={dataPanelStyles.tabPane}
-              mountOnEnter
-              unmountOnExit
-            >
-              <div className="text-info">
-                <FontAwesomeIcon icon={faInfoCircle} /> This tab is only visible
-                to developers
-              </div>
-              <DataTabJsonTree
-                data={formState ?? {}}
-                searchable
-                tabKey={DataPanelTabKey.Formik}
-                label="Formik State"
-              />
-            </Tab.Pane>
-            <Tab.Pane
-              eventKey={DataPanelTabKey.BlockConfig}
-              className={dataPanelStyles.tabPane}
-              mountOnEnter
-              unmountOnExit
-            >
-              <div className="text-info">
-                <FontAwesomeIcon icon={faInfoCircle} /> This tab is only visible
-                to developers
-              </div>
-              <DataTabJsonTree
-                data={extensionPoint}
-                tabKey={DataPanelTabKey.BlockConfig}
-                label="Configuration"
-              />
-            </Tab.Pane>
+            <StateTab />
+            <ConfigurationTab config={extensionPoint} />
           </>
         )}
         <Tab.Pane
